requirements.txt
colorama==0.3.9
Django==1.11.6
django-filter==1.1.0
djangorestframework==3.7.1
httpie==0.9.9
Markdown==2.6.9
mysqlclient==1.3.12
olefile==0.44
Pillow==4.3.0
Pygments==2.2.0
PyMySQL==0.7.11
pytz==2017.2
requests==2.13.0
six==1.11.0
transliterate==1.10
